Roster.as, RosterExtension.as:
Fixed bug where adding item to roster with null group would crash Flash Player 9.0.124

This is because Roster was always putting groupName into an array without checking to see if it was null first. Then, RosterExtension.as would automatically loop through the array and add all the elements into text nodes - so a text node would have a null value.

Then, in HTTPService.send, the Flex SDK calls XML(data) to convert it into an XMLDocument. XML() is a native function in the SDK and crashes the 9.0.124 runtime because of the null text node.